requirements.txt
Django>=3.2.4,<5.0.1
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.8.6,<2.9.9
drf-spectacular>=0.15.1,<0.27.1
pillow>=9.0.0,<10.1.0